Post, Comment report system - delete, reject report

+++ app/Livewire/Admin/UserReport.php
@@ -2,23 +2,41 @@
 
 namespace App\Livewire\Admin;
 
+use App\Models\ReportsComment;
+use App\Models\ReportsPost;
 use App\Models\ReportsUser;
 use Livewire\Component;
 
 class UserReport extends Component
 {
     public $reportId;
-    public function mount($reportId)
+    public $type;
+    public function mount($reportId, $type)
     {
         $this->reportId = $reportId;
+        $this->type = $type;
     }
     public function reject()
     {
-        $reject = ReportsUser::where('id', $this->reportId)->first();
-        
-        if($reject){
-            $reject->delete();
+        if($this->type == 'user'){
+            $reject = ReportsUser::where('id', $this->reportId)->first();
+            if($reject){
+                $reject->delete();
+            }
         }
+        if($this->type == 'post'){
+            $reject = ReportsPost::where('id', $this->reportId)->first();
+            if($reject){
+                $reject->delete();
+            }
+        }
+        if($this->type == 'comment'){
+            $reject = ReportsComment::where('id', $this->reportId)->first();
+            if($reject){
+                $reject->delete();
+            }
+        }
+        
         return redirect(request()->header('Referer'));
     }
     public function render()
